requirements.txt
streamlit==1.29.0
google-generativeai==0.8.3
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
plotly==5.17.0
pandas==2.1.4
